+++ R/DataBackendCbind.R
@@ -1,7 +1,6 @@
 DataBackendCbind = R6Class("DataBackendCbind", inherit = DataBackend, cloneable = FALSE,
   public = list(
     initialize = function(b1, b2) {
-
       assert_backend(b1)
       assert_backend(b2)
       pk = b1$primary_key
@@ -17,6 +16,7 @@ DataBackendCbind = R6Class("DataBackendCbind", inherit = DataBackend, cloneable 
 
       super$initialize(list(b1 = b1, b2 = b2), pk, "data.table")
     },
+
     data = function(rows, cols, data_format = self$data_formats[1L]) {
 
       pk = self$primary_key